started making a playstation 1 emulator
the idea is to start off with an interpreter which has a struct for the R3000
cpu, memory, CD contents, etc. then move on to a dynarec that can reuse these
structs. This differs from gb-emu in that the contents of the cpu (registers,
flags, etc.) are separated from the main functionality of the emulator (the
methods in the interpreter/future dynarec structs).