added logic to do delayed writes and made delayed jumps more idiomatic
in MIPS processors the instruction after every jump is always executed before
the program counter is updated. In a similar fashion, memory transactions are
also not available to the next opcode. I've implemented delayed stores but not
delayed loads, but now I'm not entirely sure if stores are actually delayed.