started adding graphics transformation engine (GTE) and compressed COPn macros
the idea is that the two coprocessors that the PSX implements have identical
methods which are called by the two uses of cop!(). the implementations of cop0
and gte (cop2) handle the differences in the coprocessors. Fortunately the one
command that cop0 implements doesn't modify CPU registers directly so this
design should work