replaced extra variable in memory with HashMap
Memory access to any sections that aren't explicitly handled by the various
arrays are now handled with a hashmap. technically this will only work correctly
for 32 bit reads and writes but it might be fine for the BIOS logo loop. I'll
use this solution until the gpu and gte are working then fix this kludge