fixed ORI opcode that was causing bad memory access
the problem was that immediate or and immediate add were using the same opcode
macro while add sign extends the immediate 16-bit data but or doesn't. I should
mark each opcode macro with the instructions that use it. this removed the bad
memory access and the need to use a hashmap.