implemented DMA triggering
when a memory write triggers an interrupt that method returns Some(transfers)
where transfers is a vector of DMA Transfer structs which should happen. Since
DMA transfers data between memory and peripherals, I removed the DMA struct and
redesigned it so the transfer itself happens in the interpreter. The downside to
this design is that the DMA transfer will have to be handled at each memory
write, although thanks to macros there's only one spot instead of three. All
that's left at this point is to handle the memory transfer itself in
interpreter.rs