implemented LWL and LWR
the new mov macros leave a lot to be desired in terms of clarity, but it's good
enough to continue debugging the GPU. Also (kinda important) these are
technically not correct since we're reading the value of rt from r3000 instead
of from the delay slot. This means that pairs of consecutive LWL/LWR
instructions will NOT work as expected; only the second instruction will have
an effect. I could fix this by scanning delayed_writes instead of reading rt
from r3000. This would likely be a performance hit, but since doing LWL/LWR also
had performance implications on the actual playstation, I'm assuming it isn't
used widely enough to strongly impact the emulator's performance relative to
other issues