started implementing parameter passing to CD
since it's possible to pass both a command and a parameter to the CD in one
write, it may be necessary to replace all occurences of MemAction with
Vec<MemAction> and modify resolve_memaction accordingly. This will probably be a
big performance hit, but it might be ok. If something similar arises for
MemResponse, I should almost certainly NOT do the same since reads are so much
more common