resolve_memaction now takes an Option<Vec<MemAction>> as argument
also started working on writing to the CD I/O register. It turns out that even
though only one aligned address is used as an I/O port for the CDROM controller,
the effect of writing to each byte in this word varies based on the lowest 2
bits of the first byte. In other words, the CDROM controller I/O port is
actually more complicated than the GPU I/O port...