started implementing interrupts in the memory and interpreter modules
an interrupt triggered in ioports.rs gets passed to the interpreter which calls
some cop0 method requesting an interrupt. I'm not sure if the new cop0 method is
aware of which interrupt was requested or not, but if it's not then I can
simplify the Interrupt variant of MemAction and make the loop that fills the irq
buffer short-circuiting. On a somewhat related note, I probably should remove
the need to return a Vec<MemAction> in memory::write_*() since the only time
this is used so far is when both a command and parameter are passed to the CDROM
controller.