trying different cache invalidation ideas
I don't need to do a cache invalidation unless the block I want to execute has
been invalidated. So before executing each block I'll check if that's the case
and if it is I'll do the expensive cache invalidation. However checking if I
never empty the Vec of overwritten addresses, checking this condition becomes
more expensive as time goes on. Fortunately writes tend to repeat addresses, so
instead of using a Vec to store every write I could use a set to maintain a list
of unique addresses which have been invalidated.