massive reorganization
Removed the Runnable trait and did a massive reorganization of the interpreter
and jit. The idea is that Console represents the state of the PS1 while the
struct that contains Console is in charge of emulating the console. In the case
of the interpreter this simply means having run(), step() and execute_opcode()
and console as the only member of Interpreter. In the case of the JIT the cache
becomes a member of JIT along with a Console. Then JIT defines methods to run,
compile individual opcodes, translate blocks of code, invalidate the cache and
so on. This design removes the need to duplicate code in each JIT/interpreter
implementation. I could further reduce the need for duplication by using
templates to create different variants of JIT that execute blocks in various
formats e.g. the dummy JIT sees Vec<Box<dyn Fn(&mut Console)>> as a block while
a real JIT that executes x86 machine code would see an array of memory as a
block. In this case, the generic JIT would just contain run(), compile_stub()
and invalidate_cache() as methods and the specialized JIT would have to provide
compile_opcode().