started adding more aggressive optimizations to the JIT
for now the optimized JIT just uses constant propagation, but in many cases it
doesn't actually reduce the number of operations meaning that there isn't much
of an improvement. When LUI preceeds another immediate ALU opcode the LUI is
deleted, but I could probably do more than this. The optimized read/write
opcodes (i.e. calculating the address at compile time if it's known to be a
constant) did show noticeable improvements in perf so it might be worth it to
look into other types of optimizations. also added command line JIT
optimization flag.