found a decent cache invalidation strategy
the cache is invalidated after each block meaning that we don't have to check if a block is valid in the dispatcher and the list of overwritten addresses probably won't grow much larger than 1 or 2. This is much better than doing cache invalidation lazily by checking if a block is valid in the dispatcher because in that case the list of unique overwritten addresses grew very large (10,000s in 10s of seconds) and was expected to grow even larger over a few minutes.