fixed a somewhat important bug in the caching interpreter and simplified mem
when the program counter is updated in the caching interpreter, I have to make
sure to `take()` the value in next_pc so it doesn't get reused. Apparently the
emulator works well enough to display the boot-up screen with this bug but not
well enough to get to the point where it's sending commands to the CD
controller. I think that there is another bug that prevents it from sending CD
commands which is that ADD, SUB and ADDI branch to trap overflow while the
current structure of the caching interpreter prevents them from acting as
branches. In another branch, I tried treating these instructions as branches but
that is a very bad solution since they are allowed to appear consecutively
unlike jumps in MIPS I. The caching interpreter heavily relies on this
assumption so the correct solution is to put the trapping ALU opcodes (and
potentially conditional branches) inside blocks and allow control flow to be
interrupted mid-block. This is probably easiest done on the actual JIT, but that
ofc will have the disadvantage that it's mostly in asm so it would be good to
implement the correct solution in the caching interpreter at some point for
debugging purposes.