got the new caching interpreter working
the bug with `break` not being implemented in the old caching interpreter really
highlighted the fact that the two versions of the caching interpreter will
create different blocks with the newer blocks being longer in general (always?).
When I want to leave an opcode unimplemented I should add it into Insn::new
(i.e. create a blank insn tag) and add a boxed closure with todo!() in
Stub::new. That way it'll panic only if it tries to execute the given stub, but
blocks containing the opcode can still be created.