implemented one way of calling auxilary component methods from JIT function
added a demo implementation that calls CD::exec_command(). it works but I don't
like the implementation as it'll probably lead to bloated JIT functions since
the addresses of the method and the struct instance are hardcoded at the call
site. It would be good to get all the address dependencies in at compile time,
but it might make sense to reserve a few x64 registers for these addresses since
we're moving the addresses to registers before the call anyway. Then it becomes
a question of when/how do I load the addresses into the reserved registers. The
two obvious options are load them at compile time, or load them at runtime.
Assuming 5 aux components, that gives me 5 * 2 * (64 bits + 2 bytes) = 100 bytes
per block for the first option or the runtime cost loading them just in time.