fixed the seg fault with println in fn called from JIT code
callq in the JIT code effectively pushes a 64-bit word onto the stack, but it
turns out that the stack should be aligned to 16-bytes to use certain SSE
instructions which apparently println uses. The simplest workaround is to call
subq 8 %rsp before the calling the function requiring 16-byte stack alignment
then call addq 8 %rsp afterwards to make sure we pop what we previously pushed.
In this demo the subq and addq occur right before and after the println but
ideally it should occur in the JIT code.