simplified call to JITted code and added a demo for register loading
I'm trying to see if it's possible to map the 31 non-zero MIPS registers R1..R31 onto the standard x86-64 general purpose registers, but it doesn't seem possible. The best I've gotten is 30 registers by splitting the 64-bit registers in two. This would require a rolq $32 %rax to access the MIPS register in the upper half, but that should be doable if I put the most frequently used registers in the lower halves. Not sure how to handle the 31st register since %rsp is required for calls to work correctly.