added a basic demo of loading and saving registers in the JIT
At the beginning of each block the MIPS registers which will be used are loaded
from the R3000 struct to the x64 registers they'll be mapped to. At the end of
the block the registers are written back to the R3000 struct. I should reduce
some of the redundancy in MacroAssembler before moving on to implementing LUI
instead of just loading in a dummy value.