added temp fix to JIT red-zone issue and started planning for SW insn
for some reason when red-zone optimization is enabled (i.e. release mode) I can't modify %r15 or else I get a segfault. I verified this is a problem with red-zone optimization by compiling in release mode and explicitly turning off red zone optimization. For now the workaround is to use %r14 for the register loading and saving, butcompiling without rz opt is not feasible to the interpreters so I really should figure out a long-term solution to this.

Also realized that many of the movl instructions zero out the top 32-bits of the register so I may have to rethink how to handle JIT blocks that use more than 14 MIPS registers. I could stick with 2 MIPS registers per x64 register like I planned, but that might entail replacing some movl's with movq's meaning the JIT blocks would take up more memory. I could also use the stack for excess MIPS registers. I'll already have to put up to 3 MIPS registers on the stack so I might as well make good use of the infrastructure I'll add for that and use it to store up to 17 MIPS registers.