finished writing and testing push/pop emitters
also fixed a bug in emit_movq_ir. I think with the stack emitters working, it's
now appropriate to move on to mov opcodes. To verify if a test succeeded I
should stick to pushing a result onto the stack then popping it into %rax since
that seemed to work well enough and now I know the stack opcodes work correctly.