general macro_assembler module cleanup
one somewhat important note is that the two methods that swap the MIPS registers
in and out are still part of the MacroAssembler struct but they are now outside
the macro_assembler module. The idea is that writing a generic x64 JIT compiler
is a huge endeavor so I should organize it such that I could just copy over the
macro_assembler directory to another project if I need another JIT. Then I could
just add methods as necessary so the JIT will be able to interact with the rest
of the project.