renamed some stack emitters to better reflect operand size
Also put a ptr to the MIPS register array on the stack to allow use of R15
within a JIT block. Couldn't find a way to do this without increasing the size
of the JIT function made from the first few insn in the BIOS. Ideally with
normal-sized JIT functions this small increase will be offset by the performance
gains of being able to use another register for MIPS registers. I probably
should implement movl with the SIB byte at some point, though that doesn't seem
like it'll help with the prologue/epilogue size problem.