fixed segfault in fn call in SW opcode by using sys V calling convention ABI
I now need to think about how granular the methods in macro assembler should be.
In other words, will there ever be a case where I want to use callq without
pushing caller saved registers onto the stack and fixing alignment. Maybe in the
case that one JIT fn is unconditionally followed by another I could implement it
as a callq to another JIT fn, but there might be some issues/other conditions
for this to work so I'll have to think about it.