added rax to caller_saved_registers
I should probably make some wrappers for callq_r64 to handle the no return value
case (push and pop all of caller_saved_registers) and the case where we return a
value. It would be nice if the wrapper could select where to put the return
value to give the caller less to think about.