implemented emitter for ADDIU and SLL when dest is R0
one slight issue I realized may be a problem is that mips_to_x64() gives a
runtime error when trying to get R0. One the one hand this will force me to
optimize cases where dest is R0. On the other hand, if I forget to optimize a
case where mips_to_x64() is called, it'll cause runtime errors but may not be
immediately obvious if that particular insn isn't called with dest R0. now it
looks like it's time to start thinking about how the program counter will be
updated between blocks.