implemented first two BIOS opcodes in x64 JIT
also decided to change representation of R3000 registers to match the memory
layout that jam expects. The other option was keeping R1 as the first element of
the array and subtracting 1 from the index every time we refer to Rn (i.e. every
time we call get_r[t,s,d] or get handles to the emu reg JITValues).