fixed branch delay slots in x64 JIT by fixing typo in jam
this currently doesn't work for consecutive branch delay slots. Sidenote:
consecutive jumps technically aren't valid in MIPS code, but since operations
that trap overflow effectively have branch delay slots, something like BNE
followed by ADDI is two consecutive branches. Coincidentally this is the first
conditional branch in the PS1 BIOS. the problem is that we can't save_flags
for the second branch if the flags for the first branch haven't been loaded.
I should try to find a way to do this with the existing jam Recompiler methods
or modify JITValue::Flags to act as a queue of values.